+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# 0.2.7
+- bug fix to `Doc.getModeAt`
+- added `Doc.getModeName()` and `Doc.getModeNameAt()` methods
+
 # 0.2.6
 - patch in a completion fix for https://github.com/codemirror/CodeMirror/issues/3189